network, cache: move the network cache files to a common pkg
