+++ pkg/virt-handler/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "//pkg/ephemeral-disk-utils:go_default_library",
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
         "//pkg/host-disk:go_default_library",
+        "//pkg/network/cache:go_default_library",
         "//pkg/network/driver:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/migrations:go_default_library",
@@ -25,7 +26,6 @@ go_library(
         "//pkg/virt-handler/migration-proxy:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-launcher/virtwrap/network:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network/cache:go_default_library",
         "//pkg/watchdog:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
@@ -53,6 +53,8 @@ go_test(
         "//pkg/certificates:go_default_library",
         "//pkg/ephemeral-disk-utils:go_default_library",
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
+        "//pkg/network/cache:go_default_library",
+        "//pkg/network/cache/fake:go_default_library",
         "//pkg/network/driver:go_default_library",
         "//pkg/testutils:go_default_library",
         "//pkg/virt-config:go_default_library",
@@ -65,8 +67,6 @@ go_test(
         "//pkg/virt-handler/notify-server:go_default_library",
         "//pkg/virt-launcher/notify-client:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network/cache:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network/cache/fake:go_default_library",
         "//pkg/watchdog:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
