+++ pkg/virt-handler/vm_test.go
@@ -31,11 +31,11 @@ import (
 	"sync"
 	"time"
 
+	netcache "kubevirt.io/kubevirt/pkg/network/cache"
+	fakenetcache "kubevirt.io/kubevirt/pkg/network/cache/fake"
 	netdriver "kubevirt.io/kubevirt/pkg/network/driver"
 	container_disk "kubevirt.io/kubevirt/pkg/virt-handler/container-disk"
 	hotplug_volume "kubevirt.io/kubevirt/pkg/virt-handler/hotplug-disk"
-	cache2 "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network/cache"
-	networkingfake "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network/cache/fake"
 
 	"github.com/golang/mock/gomock"
 	. "github.com/onsi/ginkgo"
@@ -199,7 +199,7 @@ var _ = Describe("VirtualMachineInstance", func() {
 			migrationProxy,
 		)
 		controller.hotplugVolumeMounter = mockHotplugVolumeMounter
-		controller.networkCacheStoreFactory = networkingfake.NewFakeInMemoryNetworkCacheFactory()
+		controller.networkCacheStoreFactory = fakenetcache.NewFakeInMemoryNetworkCacheFactory()
 		controller.virtLauncherFSRunDirPattern = filepath.Join(shareDir, "%d")
 
 		vmiTestUUID = uuid.NewUUID()
@@ -2264,7 +2264,7 @@ var _ = Describe("VirtualMachineInstance", func() {
 			vmi.Status.Phase = v1.Scheduled
 			vmi.Status.Interfaces = make([]v1.VirtualMachineInstanceNetworkInterface, 0)
 
-			podCacheInterface := &cache2.PodCacheInterface{
+			podCacheInterface := &netcache.PodCacheInterface{
 				Iface: &v1.Interface{
 					Name: interfaceName,
 				},
@@ -2321,7 +2321,7 @@ var _ = Describe("VirtualMachineInstance", func() {
 				},
 			}
 
-			podCacheInterface := &cache2.PodCacheInterface{
+			podCacheInterface := &netcache.PodCacheInterface{
 				Iface: &v1.Interface{
 					Name: interfaceName,
 				},
