+++ pkg/virt-launcher/virtwrap/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "//pkg/hooks:go_default_library",
         "//pkg/host-disk:go_default_library",
         "//pkg/ignition:go_default_library",
+        "//pkg/network/cache:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/net/ip:go_default_library",
         "//pkg/virt-handler/cmd-client:go_default_library",
@@ -33,7 +34,6 @@ go_library(
         "//pkg/virt-launcher/virtwrap/device/sriov:go_default_library",
         "//pkg/virt-launcher/virtwrap/errors:go_default_library",
         "//pkg/virt-launcher/virtwrap/network:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network/cache:go_default_library",
         "//pkg/virt-launcher/virtwrap/stats:go_default_library",
         "//pkg/virt-launcher/virtwrap/util:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
