+++ pkg/virt-launcher/virtwrap/network/BUILD.bazel
@@ -9,10 +9,10 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/network/cache:go_default_library",
         "//pkg/network/driver:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network/cache:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//staging/src/kubevirt.io/client-go/precond:go_default_library",
@@ -31,10 +31,10 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
+        "//pkg/network/cache:go_default_library",
+        "//pkg/network/cache/fake:go_default_library",
         "//pkg/network/driver:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network/cache:go_default_library",
-        "//pkg/virt-launcher/virtwrap/network/cache/fake:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/coreos/go-iptables/iptables:go_default_library",
